@@ -3,7 +3,7 @@ description: Demonstrates how to use the pytorch_mobile plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -12,8 +12,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  image_picker: ^0.6.5
-  path_provider: ^1.6.27  
+  image_picker: ^0.7.5+3
+  path_provider: ^2.0.2  
 
 dev_dependencies:
   flutter_test:
