@@ -1,17 +1,17 @@
 name: pytorch_mobile
 description: A flutter plugin for pytorch model inference. Supports image models as well as custom models.
-version: 0.1.0
+version: 0.2.0
 homepage: https://github.com/fynnmaarten/flutter_pytorch_mobile
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0 < 2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  path_provider: ^1.6.5
-  path: ^1.6.4
+  path_provider: ^2.0.2
+  path: ^1.8.0
 
 dev_dependencies:
   flutter_test:
