@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:image_picker/image_picker.dart';
-//import 'package:path_provider/path_provider.dart';
 
 import 'package:pytorch_mobile/pytorch_mobile.dart';
 import 'package:pytorch_mobile/model.dart';
@@ -17,11 +16,12 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  Model _imageModel, _customModel;
+  Model? _imageModel, _customModel;
 
-  String _imagePrediction;
-  List _prediction;
-  File _image;
+  String? _imagePrediction;
+  List? _prediction;
+  File? _image;
+  ImagePicker _picker = ImagePicker();
 
   @override
   void initState() {
@@ -44,27 +44,23 @@ class _MyAppState extends State<MyApp> {
   //run an image model
   Future runImageModel() async {
     //pick a random image
-    /*var bytes = await rootBundle.load("assets/kitten.jpeg");
-    String tempPath = (await getTemporaryDirectory()).path;
-    File image = File('$tempPath/kitten.jpeg');
-    await image.writeAsBytes(bytes.buffer.asUint8List(bytes.offsetInBytes, bytes.lengthInBytes));*/
-    File image = await ImagePicker.pickImage(
+    final PickedFile? image = await _picker.getImage(
         source: (Platform.isIOS ? ImageSource.gallery : ImageSource.camera),
         maxHeight: 224,
         maxWidth: 224);
     //get prediction
     //labels are 1000 random english words for show purposes
-    _imagePrediction = await _imageModel.getImagePrediction(
-        image, 224, 224, "assets/labels/labels.csv");
+    _imagePrediction = await _imageModel!.getImagePrediction(
+        File(image!.path), 224, 224, "assets/labels/labels.csv");
 
     setState(() {
-      _image = image;
+      _image = File(image.path);
     });
   }
 
   //run a custom model with number inputs
   Future runCustomModel() async {
-    _prediction = await _customModel
+    _prediction = await _customModel!
         .getPrediction([1, 2, 3, 4], [1, 2, 2], DType.float32);
 
     setState(() {});
@@ -80,7 +76,7 @@ class _MyAppState extends State<MyApp> {
         body: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            _image == null ? Text('No image selected.') : Image.file(_image),
+            _image == null ? Text('No image selected.') : Image.file(_image!),
             Center(
               child: Visibility(
                 visible: _imagePrediction != null,
@@ -88,14 +84,19 @@ class _MyAppState extends State<MyApp> {
               ),
             ),
             Center(
-              child: FlatButton(
+              child: TextButton(
                 onPressed: runImageModel,
-                child: Icon(Icons.add_a_photo),
+                child: Icon(
+                  Icons.add_a_photo,
+                  color: Colors.grey,
+                ),
               ),
             ),
-            FlatButton(
+            TextButton(
               onPressed: runCustomModel,
-              color: Colors.blue,
+              style: TextButton.styleFrom(
+                backgroundColor: Colors.blue,
+              ),
               child: Text(
                 "Run custom model",
                 style: TextStyle(
@@ -106,7 +107,7 @@ class _MyAppState extends State<MyApp> {
             Center(
               child: Visibility(
                 visible: _prediction != null,
-                child: Text(_prediction != null ? "${_prediction[0]}" : ""),
+                child: Text(_prediction != null ? "${_prediction![0]}" : ""),
               ),
             )
           ],
