@@ -11,9 +11,9 @@ class Model {
   Model(this._index);
 
   ///predicts abstract number input
-  Future<List> getPrediction(
+  Future<List?> getPrediction(
       List<double> input, List<int> shape, DType dtype) async {
-    final List prediction = await _channel.invokeListMethod('predict', {
+    final List? prediction = await _channel.invokeListMethod('predict', {
       "index": _index,
       "data": input,
       "shape": shape,
@@ -27,7 +27,7 @@ class Model {
       File image, int width, int height, String labelPath) async {
     List<String> labels = await _getLabels(labelPath);
     List byteArray = image.readAsBytesSync();
-    final List prediction = await _channel.invokeListMethod("predictImage", {
+    final List? prediction = await _channel.invokeListMethod("predictImage", {
       "index": _index,
       "image": byteArray,
       "width": width,
@@ -35,7 +35,7 @@ class Model {
     });
     double maxScore = double.negativeInfinity;
     int maxScoreIndex = -1;
-    for (int i = 0; i < prediction.length; i++) {
+    for (int i = 0; i < prediction!.length; i++) {
       if (prediction[i] > maxScore) {
         maxScore = prediction[i];
         maxScoreIndex = i;
@@ -45,9 +45,14 @@ class Model {
   }
 
   ///predicts image but returns the raw net output
-  Future<List> getImagePredictionList(File image, int width, int height) async {
-    final List prediction = await _channel.invokeListMethod("predictImage",
-        {"index": _index, "image": image.readAsBytesSync(), "width": width, "height": height});
+  Future<List?> getImagePredictionList(
+      File image, int width, int height) async {
+    final List? prediction = await _channel.invokeListMethod("predictImage", {
+      "index": _index,
+      "image": image.readAsBytesSync(),
+      "width": width,
+      "height": height
+    });
     return prediction;
   }
 
